feat(room view): add `disableReadReceiptsAndMarkersOnActivity` props
For the multiroom module, we display several room views at the same
time. In order to avoid all the rooms to send read receipts and markers
automatically when we are interacting with the UI, we add
`disableReadReceiptsAndMarkersOnActivity`props.

When at true, the timeline doesn't listen to user activity to send these
receipts. Only when the room is focused, marker and read receipts are
updated.